Implement Chat component functionality
Started out mocking the useAppState and useAppDis-
patch hooks, but this ended up being very clunky.

Moved away from isolating the component, but isol-
ating the test case instead. In this case, the re-
sulting tests are much more readable and useful.

Simpler and with no mocking which ended up being
basically just a total re-implementation of the
existing code. May as well use the real thing.

The tests now also check the end result, not the
implementation, which is something I keep stumbli-
ng into.